+++ t2/settings/settings.h
@@ -6,7 +6,7 @@ namespace t2 {
 		extern float camera_move_speed_z;
 		extern float camera_rotation_speed_pitch;
 		extern float camera_rotation_speed_yaw;
-		extern float camera_fov;
+		extern int camera_fov;
 		//extern bool set_camera;
 		extern bool show_player_model;
 		extern bool show_weapon_model;
